Markuse asjade juurutamise tööriist alpha2 + Toimiv sõrmejälje genereerimine + Seadistusaken avaneb, kui õige sõrmejäljega mälupulk on sisestatud

+++ Markuse asjade juurutamise tööriist/Form1.cs
@@ -9,6 +9,7 @@ using System.Windows.Forms;
 using System.Management;
 using System.Security.Cryptography;
 using System.IO;
+using Microsoft.Win32;
 
 namespace Markuse_asjade_juurutamise_tööriist
 {
@@ -27,6 +28,13 @@ namespace Markuse_asjade_juurutamise_tööriist
         
         private void Form1_Load(object sender, EventArgs e)
         {
+            RegistryKey wineKey = Registry.LocalMachine.OpenSubKey(@"Software\Wine", false);
+            if (wineKey != null)
+            {
+                MessageBox.Show("Markuse arvuti asjade juurutamistarkvara ei käivitu Wine all. Palun käivitage see programm Windows-is.");
+                return;
+            }
+
             if (this.Width < 574)
             {
                 label1.Visible = false;
@@ -171,6 +179,56 @@ namespace Markuse_asjade_juurutamise_tööriist
                 pictureBox2.Image = Properties.Resources.root_idc_key;
                 label3.Text = "Andmete dekrüptimine...";
             }
+            else if (action == "decrypt")
+            {
+                pictureBox2.Image = null;
+                label3.Text = "";
+                if (!File.Exists(flashID + "\\E_INFO\\ROOT_INF\\verify.dat"))
+                {
+                    string[] letters = { "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z" };
+                    foreach (string letter in letters)
+                    {
+                        if (letter + ":" != Environment.GetEnvironmentVariable("HOMEDRIVE"))
+                        {
+                            if (File.Exists(letter + ":\\E_INFO\\ROOT_INF\\verify.dat"))
+                            {
+                                flashID = letter + ":";
+                            }
+                        }
+                    }
+                }
+                if (File.Exists(flashID + "\\E_INFO\\edition.cmd")) {
+                    string pass_phrase = File.ReadAllText(flashID + "\\E_INFO\\edition.cmd").ToLower();
+                    string flash_serial = GetSerialFromDrive(flashID);
+                    string verifyData = File.ReadAllText(flashID + "\\E_INFO\\ROOT_INF\\verify.dat", Encoding.BigEndianUnicode);
+                    actionWaitTimer.Enabled = false;
+                    string claimed_serial = StringCipher.Decrypt(verifyData, pass_phrase);
+                    if (claimed_serial != flash_serial)
+                    {
+                        MessageBox.Show("Sõrmejälg on vale. Võimalik, et kopeerisite verify.dat faili ühelt mälupulgalt teisele ning ei uuendanud sõrmejälge. Programm sulgub nüüd.\n\n" + claimed_serial + " != " + flash_serial, "Markuse asjade juurutamise tööriist", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        canClose = true;
+                        this.Close();
+                    }
+                } else
+                {
+                    actionWaitTimer.Enabled = false;
+                    MessageBox.Show("Sõrmejäljega Markuse mälupulka ei leitud. Programm sulgub nüüd.", "Markuse asjade juurutamise tööriist peab sulguma", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    canClose = true;
+                    this.Close();
+                }
+                actionWaitTimer.Enabled = false;
+                label5.Text = "Etapp 2: Juurutamine";
+                this.Hide();
+                Setup setup = new Setup();
+                if (setup.ShowDialog() == DialogResult.OK)
+                {
+                    this.Show();
+                } else
+                {
+                    canClose = true;
+                    this.Close();
+                }
+            }
             else if (action == "flash_w")
             {
                 if (!Directory.Exists(flashID + "\\E_INFO\\ROOT_INF"))
@@ -179,19 +237,28 @@ namespace Markuse_asjade_juurutamise_tööriist
                 } else
                 {
                     actionWaitTimer.Enabled = false;
-                    if (MessageBox.Show("Sõrmejälg on juba eelnevalt loodud. Kas soovite selle üle kirjutada?", "Markuse asjade juurutamise tööriist", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation) == DialogResult.Yes)
-                    { 
-                        Directory.Delete(flashID + "\\E_INFO\\ROOT_INF", true);
-                        Directory.CreateDirectory(flashID + "\\E_INFO\\ROOT_INF");
-                    } else
+                    if (File.Exists(flashID + "\\E_INBFO\\ROOT_INF\\verify.dat"))
                     {
-                        MessageBox.Show("Juurutamise tööriist sulgub, kuna tegum nurjus. P: Sõrmejälge ei saa luua", "Markuse asjade juurutamise tööriist peab sulguma", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                        canClose = true;
-                        this.Close();
+                        if (MessageBox.Show("Sõrmejälg on juba eelnevalt loodud. Kas soovite selle üle kirjutada?", "Markuse asjade juurutamise tööriist", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation) == DialogResult.Yes)
+                        {
+                            Directory.Delete(flashID + "\\E_INFO\\ROOT_INF", true);
+                            Directory.CreateDirectory(flashID + "\\E_INFO\\ROOT_INF");
+                        }
+                        else
+                        {
+                            MessageBox.Show("Juurutamise tööriist sulgub, kuna tegum nurjus. P: Sõrmejälge ei saa luua", "Markuse asjade juurutamise tööriist peab sulguma", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            canClose = true;
+                            this.Close();
+                        }
                     }
-                    MessageBox.Show(GetSerialFromDrive(flashID));
-                    action = "end";
-                    actionWaitTimer.Enabled = true;
+                    string pass_phrase = File.ReadAllText(flashID + "\\E_INFO\\edition.cmd", Encoding.ASCII).ToLower();
+                    string unique_code = StringCipher.Encrypt(GetSerialFromDrive(flashID), pass_phrase);
+                    File.WriteAllText(flashID + "\\E_INFO\\ROOT_INF\\verify.dat", unique_code, Encoding.BigEndianUnicode);
+                    MessageBox.Show("Sõrmejälg loodi edukalt " + flashID + " kettale.", "Markuse asjade juurutamise tööriist", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    canClose = true;
+                    this.Close();
+                    // action = "end";
+                    // actionWaitTimer.Enabled = true;
                 }
             }
         }
@@ -376,6 +443,101 @@ namespace Markuse_asjade_juurutamise_tööriist
             } else if (canClose == true)
             {
                 e.Cancel = false;
+            }
+        }
+
+        // krüpti/dekrüpti
+        // https://stackoverflow.com/questions/10168240/encrypting-decrypting-a-string-in-c-sharp
+        public static class StringCipher
+        {
+            // This constant is used to determine the keysize of the encryption algorithm in bits.
+            // We divide this by 8 within the code below to get the equivalent number of bytes.
+            private const int Keysize = 256;
+
+            // This constant determines the number of iterations for the password bytes generation function.
+            private const int DerivationIterations = 1445;
+
+            public static string Encrypt(string plainText, string passPhrase)
+            {
+                // Salt and IV is randomly generated each time, but is preprended to encrypted cipher text
+                // so that the same Salt and IV values can be used when decrypting.  
+                var saltStringBytes = Generate256BitsOfRandomEntropy();
+                var ivStringBytes = Generate256BitsOfRandomEntropy();
+                var plainTextBytes = Encoding.UTF8.GetBytes(plainText);
+                using (var password = new Rfc2898DeriveBytes(passPhrase, saltStringBytes, DerivationIterations))
+                {
+                    var keyBytes = password.GetBytes(Keysize / 8);
+                    using (var symmetricKey = new RijndaelManaged())
+                    {
+                        symmetricKey.BlockSize = 256;
+                        symmetricKey.Mode = CipherMode.CBC;
+                        symmetricKey.Padding = PaddingMode.ISO10126;
+                        using (var encryptor = symmetricKey.CreateEncryptor(keyBytes, ivStringBytes))
+                        {
+                            using (var memoryStream = new MemoryStream())
+                            {
+                                using (var cryptoStream = new CryptoStream(memoryStream, encryptor, CryptoStreamMode.Write))
+                                {
+                                    cryptoStream.Write(plainTextBytes, 0, plainTextBytes.Length);
+                                    cryptoStream.FlushFinalBlock();
+                                    // Create the final bytes as a concatenation of the random salt bytes, the random iv bytes and the cipher bytes.
+                                    var cipherTextBytes = saltStringBytes;
+                                    cipherTextBytes = cipherTextBytes.Concat(ivStringBytes).ToArray();
+                                    cipherTextBytes = cipherTextBytes.Concat(memoryStream.ToArray()).ToArray();
+                                    memoryStream.Close();
+                                    cryptoStream.Close();
+                                    return Convert.ToBase64String(cipherTextBytes);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            public static string Decrypt(string cipherText, string passPhrase)
+            {
+                // Get the complete stream of bytes that represent:
+                // [32 bytes of Salt] + [32 bytes of IV] + [n bytes of CipherText]
+                var cipherTextBytesWithSaltAndIv = Convert.FromBase64String(cipherText);
+                // Get the saltbytes by extracting the first 32 bytes from the supplied cipherText bytes.
+                var saltStringBytes = cipherTextBytesWithSaltAndIv.Take(Keysize / 8).ToArray();
+                // Get the IV bytes by extracting the next 32 bytes from the supplied cipherText bytes.
+                var ivStringBytes = cipherTextBytesWithSaltAndIv.Skip(Keysize / 8).Take(Keysize / 8).ToArray();
+                // Get the actual cipher text bytes by removing the first 64 bytes from the cipherText string.
+                var cipherTextBytes = cipherTextBytesWithSaltAndIv.Skip((Keysize / 8) * 2).Take(cipherTextBytesWithSaltAndIv.Length - ((Keysize / 8) * 2)).ToArray();
+
+                using (var password = new Rfc2898DeriveBytes(passPhrase, saltStringBytes, DerivationIterations))
+                {
+                    var keyBytes = password.GetBytes(Keysize / 8);
+                    using (var symmetricKey = new RijndaelManaged())
+                    {
+                        symmetricKey.BlockSize = 256;
+                        symmetricKey.Mode = CipherMode.CBC;
+                        symmetricKey.Padding = PaddingMode.ISO10126;
+                        using (var decryptor = symmetricKey.CreateDecryptor(keyBytes, ivStringBytes))
+                        {
+                            using (var memoryStream = new MemoryStream(cipherTextBytes))
+                            {
+                                using (var cryptoStream = new CryptoStream(memoryStream, decryptor, CryptoStreamMode.Read))
+                                using (var streamReader = new StreamReader(cryptoStream, Encoding.UTF8))
+                                {
+                                    return streamReader.ReadToEnd();
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+
+            private static byte[] Generate256BitsOfRandomEntropy()
+            {
+                var randomBytes = new byte[32]; // 32 Bytes will give us 256 bits.
+                using (var rngCsp = new RNGCryptoServiceProvider())
+                {
+                    // Fill the array with cryptographically secure random bytes.
+                    rngCsp.GetBytes(randomBytes);
+                }
+                return randomBytes;
             }
         }
     }
